requirements.txt
google-generativeai
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
bcrypt==4.2.1
blinker==1.9.0
blis==1.1.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
cryptography==44.0.0
cymem==2.0.10
deprecation==2.1.0
docx2txt==0.8
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
gotrue==2.11.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
murmurhash==1.0.11
narwhals==1.19.1
nltk==3.9.1
numpy==2.2.1
packaging==24.2
pafy==0.5.5
pandas==2.2.3
pdfminer.six==20240706
pdfminer3==2018.12.3.0
pillow==11.0.0
plotly==5.24.1
postgrest==0.18.0
preshed==3.0.9
propcache==0.2.1
protobuf==5.29.2
pyarrow==18.1.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
pyresparser==1.0.6
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
realtime==2.0.6
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.0
storage3==0.9.0
streamlit==1.41.1
streamlit-tags==1.2.8
supabase==2.10.0
supafunc==0.7.0
tenacity==9.0.0
thinc==8.3.3
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
websockets==13.1
wrapt==1.17.0
yarl==1.18.3
